+++ Timelog/Views/Home/StopSessionSheet.swift
@@ -84,6 +84,8 @@ struct StopSessionSheet: View {
                 }
             }
         }
+        .onAppear    { RestSyncService.shared.isUserEditing = true  }
+        .onDisappear { RestSyncService.shared.isUserEditing = false }
     }
 
     private func addLabel(to project: Project) {